refactor/#2: AOP test 패키지 위치 변경
