+++ src/main/java/com/umc/bios/global/test/HelloController.java
@@ -1,6 +1,5 @@
-package com.umc.bios.domain.member.controller;
+package com.umc.bios.global.test;
 
-import com.umc.bios.domain.member.service.HelloService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
