+++ src/main/java/com/umc/bios/global/test/HelloService.java
@@ -1,0 +1,11 @@
+package com.umc.bios.global.test;
+
+import org.springframework.stereotype.Service;
+
+@Service
+public class HelloService {
+
+    public String sayHello() {
+        return "Hello World";
+    }
+}
